move testing utilities to better location
